Allow choosing between version.dll or uxtheme.dll. (#7)

+++ src/_unpatch.py
@@ -48,7 +48,7 @@ def revert_assets():
             os.remove(asset_backup)
 
 
-def revert_assembly(dl_latest=False, dll_name='version.dll'):
+def revert_assembly(dl_latest=False):
     print("Reverting translations.txt")
     game_folder = util.get_game_folder()
 
@@ -61,7 +61,6 @@ def revert_assembly(dl_latest=False, dll_name='version.dll'):
         return
 
     translations_path = os.path.join(game_folder, "translations.txt")
-    dll_path = os.path.join(game_folder, dll_name)
 
     if not os.path.exists(translations_path):
         print("translations.txt does not exist. Skipping.")
@@ -69,16 +68,19 @@ def revert_assembly(dl_latest=False, dll_name='version.dll'):
         os.remove(translations_path)
 
     if dl_latest:
-        if os.path.exists(dll_path):
-            print(f"Deleting {dll_name}")
-            os.remove(dll_path)
+        dll_name = settings['dll_name']
+        if dll_name:
+            settings['dll_name'] = None
+            dll_path = os.path.join(game_folder, dll_name)
 
-        bak_name = settings['dll_backup']
-        if bak_name:
-            bak_path = os.path.join(game_folder, bak_name)
+            if os.path.exists(dll_path):
+                print(f"Deleting {dll_name}")
+                os.remove(dll_path)
+
+            bak_path = os.path.join(game_folder, dll_path + util.DLL_BACKUP_SUFFIX)
             
             if os.path.exists(bak_path):
-                print(f"Restoring previous {bak_name}")
+                print(f"Restoring previous {dll_name}")
                 shutil.move(bak_path, dll_path)
 
     else:
